backend/foodgram/requirements.txt
asgiref==3.7.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.20
django-colorfield==0.9.0
django-cors-headers==4.1.0
drf-extra-fields==3.0.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
importlib-metadata==1.7.0
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2020.1
reportlab==4.0.4
requests==2.26.0
requests-oauthlib==1.3.1
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.3.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.25.11
zipp==3.8.0
